# BlockChainAppDev/PetBoarding/docs/swagger.yaml
...
openapi: 3.0.0
info:
  title: 宠物寄养系统API
  description: |
    宠物寄养系统的RESTful API文档，包含用户、宠物、寄养服务、订单、评价和通知等功能。
    
    ## 认证
    大部分API需要通过JWT令牌进行认证，请在请求头中添加`Authorization: Bearer {token}`。
    
    ## 错误处理
    所有API响应均使用统一的格式，错误时会返回相应的错误码和错误信息。
  version: 1.0.0
  contact:
    name: 宠物寄养系统团队
    email: support@petboarding.example.com
servers:
  - url: http://localhost:8080/api
    description: 本地开发环境
  - url: https://api.petboarding.example.com/api
    description: 生产环境
...
paths:
...
  /boarding:
    post:
      tags:
        - 寄养服务
      summary: 创建寄养服务
      description: 创建新的寄养服务
      operationId: createBoardingService
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoardingServiceCreateRequest'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardingServiceResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /boarding/{serviceId}:
    get:
      tags:
        - 寄养服务
      summary: 获取寄养服务信息
      description: 获取指定寄养服务的信息
      operationId: getBoardingServiceInfo
      parameters:
        - name: serviceId
          in: path
          description: 服务ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功获取寄养服务信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardingServiceResponse'
        '404':
          description: 寄养服务不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - 寄养服务
      summary: 更新寄养服务信息
      description: 更新指定寄养服务的信息
      operationId: updateBoardingServiceInfo
      parameters:
        - name: serviceId
          in: path
          description: 服务ID
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoardingServiceUpdateRequest'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardingServiceResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 寄养服务不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /notifications/user/{userId}:
    get:
      tags:
        - 通知
      summary: 获取用户的所有通知
      description: 获取指定用户的所有通知
      operationId: getUserNotifications
      parameters:
        - name: userId
          in: path
          description: 用户ID
          required: true
          schema:
            type: string
        - name: read
          in: query
          description: 是否已读，0-未读，1-已读
          required: false
          schema:
            type: integer
        - name: page
          in: query
          description: 页码，默认1
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: 每页数量，默认10
          required: false
          schema:
            type: integer
            default: 10
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功获取用户的所有通知
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationListResponse'
        '404':
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /notifications/{notificationId}:
    get:
      tags:
        - 通知
      summary: 获取通知详情
      description: 获取指定通知的详情
      operationId: getNotificationInfo
      parameters:
        - name: notificationId
          in: path
          description: 通知ID
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功获取通知详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResponse'
        '404':
          description: 通知不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /notifications/{notificationId}/read:
    put:
      tags:
        - 通知
      summary: 标记通知为已读
      description: 标记指定通知为已读
      operationId: markNotificationAsRead
      parameters:
        - name: notificationId
          in: path
          description: 通知ID
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 标记成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: 通知不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /notifications/user/{userId}/read-all:
    put:
      tags:
        - 通知
      summary: 标记所有通知为已读
      description: 标记指定用户的所有通知为已读
      operationId: markAllNotificationsAsRead
      parameters:
        - name: userId
          in: path
          description: 用户ID
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 标记成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # 管理员服务API
  /admin/login:
    post:
      tags:
        - 管理员
      summary: 管理员登录
      description: 管理员登录并获取令牌
      operationId: adminLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminLoginResponse'
        '401':
          description: 登录失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/dashboard:
    get:
      tags:
        - 管理员
      summary: 获取系统统计数据
      description: 获取系统统计数据
      operationId: getDashboardData
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功获取系统统计数据
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardResponse'
        '401':
          description: 未授权
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          nullable: true
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 400
        message:
          type: string
          example: "Invalid parameters"
        data:
          type: object
          nullable: true

    # 用户相关
...
    BoardingServiceCreateRequest:
      type: object
      required:
        - title
        - price
        - location
        - pet_type
        - capacity
      properties:
        title:
          type: string
          example: "Luxury Pet Boarding"
        price:
          type: number
          format: float
          example: 50.00
        location:
          type: string
          example: "123 Main St, City"
        pet_type:
          type: string
          example: "Dog"
        capacity:
          type: integer
          example: 5
        description:
          type: string
          example: "Luxury boarding with premium care"
        facilities:
          type: array
          items:
            type: string
          example: ["Air Conditioning", "Heated Floors", "Outdoor Play Area"]
        services:
          type: array
          items:
            type: string
          example: ["Daily Walks", "Grooming", "Training"]
        rules:
          type: string
          example: "Pets must be vaccinated and friendly"
    BoardingServiceResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          properties:
            service_id:
              type: string
              example: "s123456789"
            provider_id:
              type: string
              example: "u123456789"
            provider_name:
              type: string
              example: "John Doe"
            title:
              type: string
              example: "Luxury Pet Boarding"
            price:
              type: number
              format: float
              example: 50.00
            location:
              type: string
              example: "123 Main St, City"
            pet_type:
              type: string
              example: "Dog"
            capacity:
              type: integer
              example: 5
            available:
              type: integer
              example: 1
            rating:
              type: number
              format: float
              example: 4.8
            review_count:
              type: integer
              example: 25
            description:
              type: string
              example: "Luxury boarding with premium care"
            facilities:
              type: array
              items:
                type: string
              example: ["Air Conditioning", "Heated Floors", "Outdoor Play Area"]
            services:
              type: array
              items:
                type: string
              example: ["Daily Walks", "Grooming", "Training"]
            rules:
              type: string
              example: "Pets must be vaccinated and friendly"
            photos:
              type: array
              items:
                type: string
              example: ["url1.jpg", "url2.jpg"]
            created_at:
              type: string
              format: date-time
              example: "2023-06-01T17:00:00Z"
            updated_at:
              type: string
              format: date-time
              example: "2023-06-01T17:00:00Z"
    BoardingServiceUpdateRequest:
      type: object
      properties:
        title:
          type: string
          example: "Premium Luxury Pet Boarding"
        price:
          type: number
          format: float
          example: 55.00
        capacity:
          type: integer
          example: 6
        available:
          type: integer
          enum: [0, 1]
          example: 0
        description:
          type: string
          example: "Updated description"
...
    NotificationResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          properties:
            notification_id:
              type: string
              example: "n123456789"
            user_id:
              type: string
              example: "u123456789"
            type:
              type: integer
              example: 1
            title:
              type: string
              example: "订单状态更新"
            content:
              type: string
              example: "您的订单 #o123456789 已被接受"
            related_id:
              type: string
              example: "o123456789"
            read:
              type: integer
              enum: [0, 1]
              example: 0
            created_at:
              type: string
              format: date-time
              example: "2023-06-02T09:00:00Z"
            updated_at:
              type: string
              format: date-time
              example: "2023-06-02T09:00:00Z"
    NotificationListResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          properties:
            total:
              type: integer
              example: 20
            unread_count:
              type: integer
              example: 5
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 10
            notifications:
              type: array
              items:
                type: object
                properties:
                  notification_id:
                    type: string
                    example: "n123456789"
                  type:
                    type: integer
                    example: 1
                  title:
                    type: string
                    example: "订单状态更新"
                  content:
                    type: string
                    example: "您的订单 #o123456789 已被接受"
                  related_id:
                    type: string
                    example: "o123456789"
                  read:
                    type: integer
                    enum: [0, 1]
                    example: 0
                  created_at:
                    type: string
                    format: date-time
                    example: "2023-06-02T09:00:00Z"

    # 管理员相关
    AdminLoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: "admin"
        password:
          type: string
          example: "admin123"
    AdminLoginResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          properties:
            admin_id:
              type: string
              example: "a123456789"
            username:
              type: string
              example: "admin"
            token:
              type: string
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            expires_at:
              type: string
              format: date-time
              example: "2023-06-02T12:00:00Z"
    DashboardResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          properties:
            user_count:
              type: integer
              example: 1000
            service_count:
              type: integer
              example: 150
            order_count:
              type: integer
              example: 500
            order_stats:
              type: object
              properties:
                pending:
                  type: integer
                  example: 50
                in_progress:
                  type: integer
                  example: 100
                completed:
                  type: integer
                  example: 300
                cancelled:
                  type: integer
                  example: 50
            revenue:
              type: number
              format: float
              example: 25000.00
            recent_orders:
              type: array
              items:
                type: object
                properties:
                  order_id:
                    type: string
                    example: "o123456789"
                  user_id:
                    type: string
                    example: "u123456789"
                  service_id:
                    type: string
                    example: "s123456789"
                  total_price:
                    type: number
                    format: float
                    example: 250.00
                  status:
                    type: integer
                    example: 1
                  created_at:
                    type: string
                    format: date-time
                    example: "2023-06-01T19:00:00Z"

    # 健康检查
...